Also analyze the YAML files in the CI chain (see #1681)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

4ef19293 Also analyze the YAML files in the CI chain
1e6e1dd1 Adjust the pip install command
9fd2631a Use "python3 -m pip" instead of "pip3"